Update index.md
Hi Sam, 

I've updated a few of the acronyms so that we're doing full title and acronym first, with repeats of the same acronym including the hover text. 

I've gone with SbD as it's a little 'b' for 'by' and this is done elsewhere, and I think we need to be CIS and then use 'division' - let me know if we need to update the glossary at all as a result of those choices.

I've also just adapted the text a little to be more active voice (the few places we can) but if I've changed meaning at all (I'm 95% certain I haven't but let me know) and we can sort. Under 'Contact' I've renamed the Slack channels with their full titles. 

Let me know if I've messed anything up, ha